Added JWT Claims to Delete an item by a specific User only.

+++ 6. Flask-SQLAlchemy/resources/item.py
@@ -1,4 +1,4 @@
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import get_jwt, jwt_required
 from flask_restful import Resource, reqparse
 from models.item import ItemModel
 
@@ -44,7 +44,12 @@ class Item(Resource):
         item.save_to_db()
         return item.json()
     
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+        
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
